Refactor module layout of runtime
No functional change - just source code organization.
