+++ northstar-runtime/src/runtime/env.rs
@@ -1,0 +1,8 @@
+/// Environment variable name passed to the container with the containers name
+pub(crate) const NAME: &str = "NORTHSTAR_NAME";
+/// Environment variable name passed to the container with the containers version
+pub(crate) const VERSION: &str = "NORTHSTAR_VERSION";
+/// Environment variable name passed to the container with the containers id
+pub(crate) const CONTAINER: &str = "NORTHSTAR_CONTAINER";
+/// Environment variable name passed to the container with the console fd
+pub(crate) const CONSOLE: &str = "NORTHSTAR_CONSOLE";
